add restriction to protect db in demo mode.

+++ src/page/banner/banner_delete.go
@@ -42,6 +42,12 @@ func Delete(rw http.ResponseWriter, req *http.Request) {
 		return
 	}
 
+	if app.Ini.AppType == "demo" && !ctx.IsSuperuser() {
+		ctx.Msg.Warning(ctx.T("This function is not permitted in demo mode."))
+		ctx.Redirect(ctx.U("/banner"))
+		return
+	}
+
 	sqlStr := `delete from
 					banner
 				where
